Migrated Equation across to using the visitor
