+++ src/equations.rs
@@ -1,9 +1,4 @@
-use std::io::Write;
 use std::slice::Iter;
-
-use super::Renderable;
-use errors::*;
-
 
 /// A single equation.
 ///
@@ -34,9 +29,9 @@ use errors::*;
 /// ```
 #[derive(Clone, Debug, Default, PartialEq)]
 pub struct Equation {
-    text: String,
-    label: Option<String>,
-    not_numbered: bool,
+    pub(crate) text: String,
+    pub(crate) label: Option<String>,
+    pub(crate) not_numbered: bool,
 }
 
 impl Equation {
@@ -69,24 +64,6 @@ impl Equation {
     }
 }
 
-impl Renderable for Equation {
-    fn render<W>(&self, writer: &mut W) -> Result<()>
-        where W: Write
-    {
-        write!(writer, r"{}", self.text)?;
-        if let Some(ref label) = self.label {
-            write!(writer, r" \label{{{}}}", label)?;
-        }
-        if self.not_numbered {
-            write!(writer, r" \nonumber")?;
-        }
-
-        writeln!(writer, r" \\")?;
-        Ok(())
-    }
-}
-
-
 /// A list of equations to be used in an `align` environment.
 #[derive(Clone, Debug, Default, PartialEq)]
 pub struct Align {
@@ -111,22 +88,6 @@ impl Align {
     }
 }
 
-impl Renderable for Align {
-    fn render<W>(&self, writer: &mut W) -> Result<()>
-        where W: Write
-    {
-        writeln!(writer, r"\begin{{align}}")?;
-
-        for item in &self.items {
-            item.render(writer)?;
-        }
-
-        writeln!(writer, r"\end{{align}}")?;
-        Ok(())
-    }
-}
-
-
 impl<'a> From<&'a str> for Equation {
     fn from(other: &'a str) -> Equation {
         Equation::new(other)
@@ -140,76 +101,5 @@ impl<'a> From<&'a str> for Align {
         let mut eq = Align::new();
         eq.push(other);
         eq
-    }
-}
-
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn render_empty_align() {
-        let should_be = "\\begin{align}\n\\end{align}\n";
-        let equations = Align::new();
-
-        let mut buffer = Vec::new();
-        equations.render(&mut buffer).unwrap();
-
-        assert_eq!(String::from_utf8(buffer).unwrap(), should_be);
-    }
-
-    #[test]
-    fn render_simple_equation() {
-        let should_be = "x &= y + \\sigma \\\\\n";
-        let eq = Equation::new(r"x &= y + \sigma");
-
-        let mut buffer = Vec::new();
-        eq.render(&mut buffer).unwrap();
-
-        assert_eq!(String::from_utf8(buffer).unwrap(), should_be);
-    }
-
-    #[test]
-    fn render_several_equations() {
-        let should_be = r"\begin{align}
-E &= m c^2 \label{eq:mass-energy-equivalence} \\
-y &= m x + c \\
-\end{align}
-";
-        let mut equations = Align::new();
-
-        let mut eq = Equation::new("E &= m c^2");
-        eq.label("eq:mass-energy-equivalence");
-        equations.push(eq).push("y &= m x + c");
-
-        let mut buffer = Vec::new();
-        equations.render(&mut buffer).unwrap();
-
-        assert_eq!(String::from_utf8(buffer).unwrap(), should_be);
-    }
-
-    #[test]
-    fn equation_with_label() {
-        let should_be = "E &= m c^2 \\label{eq:mass-energy-equivalence} \\\\\n";
-        let mut eq = Equation::new("E &= m c^2");
-        eq.label("eq:mass-energy-equivalence");
-
-        let mut buffer = Vec::new();
-        eq.render(&mut buffer).unwrap();
-
-        assert_eq!(String::from_utf8(buffer).unwrap(), should_be);
-    }
-
-    #[test]
-    fn equation_with_no_numbering() {
-        let should_be = "E &= m c^2 \\nonumber \\\\\n";
-        let mut eq = Equation::new("E &= m c^2");
-        eq.not_numbered();
-
-        let mut buffer = Vec::new();
-        eq.render(&mut buffer).unwrap();
-
-        assert_eq!(String::from_utf8(buffer).unwrap(), should_be);
     }
 }
